after getting my result

+++ ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: pyathams <pyathams@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 13:50:22 by pyathams          #+#    #+#             */
-/*   Updated: 2024/03/12 17:56:19 by pyathams         ###   ########.fr       */
+/*   Updated: 2024/03/12 18:13:27 by pyathams         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
